+++ app/src/main/AndroidManifest.xml
@@ -44,8 +44,13 @@
         </service>
 
         <activity
-            android:name=".FGSignIn"
+            android:name=".LoginActivity"
             android:label="@string/app_name" >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER"/>
+            </intent-filter>
         </activity>
 
         <activity
@@ -58,15 +63,6 @@
             android:name="com.facebook.sdk.ApplicationId"
             android:value="@string/app_id"/>
 
-        <activity
-            android:name=".LoginActivity"
-            android:label="@string/app_name">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER"/>
-            </intent-filter>
-        </activity>
         <activity
             android:name=".RegisterActivity"
             android:label="@string/app_name">
