+++ app/src/main/res/layout/activity_fgsign_in.xml
@@ -7,7 +7,7 @@
                 android:paddingRight="@dimen/activity_horizontal_margin"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:paddingBottom="@dimen/activity_vertical_margin"
-                tools:context="aloksharma.ufl.edu.stash.FGSignIn">
+                tools:context="aloksharma.ufl.edu.stash.LoginActivity">
 
     <TextView
         android:text="@string/hello_world"
