+++ app/src/main/res/menu/menu_fgsign_in.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="aloksharma.ufl.edu.stash.FGSignIn">
+      tools:context="aloksharma.ufl.edu.stash.LoginActivity">
     <item
         android:id="@+id/action_settings"
         android:title="@string/action_settings"
